Move app-rpc into app-toolbox
