--- packages/app-rpc/README.md
@@ -1,1 +1,0 @@
-# @polkadot/app-rpc
