--- packages/app-rpc/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@polkadot/app-rpc",
-  "version": "0.22.40",
-  "main": "index.js",
-  "repository": "https://github.com/polkadot-js/apps.git",
-  "author": "Jaco Greeff <jacogr@gmail.com>",
-  "maintainers": [
-    "Jaco Greeff <jacogr@gmail.com>"
-  ],
-  "contributors": [],
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@babel/runtime": "^7.2.0",
-    "@polkadot/ui-app": "^0.22.40",
-    "@polkadot/ui-signer": "^0.22.40"
-  }
-}
