+++ packages/app-toolbox/src/Rpc/Results.tsx
@@ -1,4 +1,4 @@
-// Copyright 2017-2019 @polkadot/app-rpc authors & contributors
+// Copyright 2017-2019 @polkadot/app-toolbox authors & contributors
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
