+++ packages/app-toolbox/src/index.css
@@ -2,10 +2,6 @@
 /* This software may be modified and distributed under the terms
 /* of the Apache-2.0 license. See the LICENSE file for details. */
 
-.toolbox--App {
-  padding: 1em 2em;
-}
-
 .toolbox--App-navigation.ui--Button-Group {
   margin-bottom: 2em;
   text-align: center;
