--- packages/apps/src/routing/rpc.ts
@@ -1,20 +1,0 @@
-// Copyright 2017-2019 @polkadot/apps authors & contributors
-// This software may be modified and distributed under the terms
-// of the Apache-2.0 license. See the LICENSE file for details.
-
-import { Routes } from '../types';
-
-import Rpc from '@polkadot/app-rpc/index';
-
-export default ([
-  {
-    Component: Rpc,
-    i18n: {
-      defaultValue: 'Raw RPC'
-    },
-    icon: 'code',
-    isApiGated: true,
-    isHidden: false,
-    name: 'rpc'
-  }
-] as Routes);
